mapper: error refactoring
This kind of hoop-jumping is unpleasant, but I'm currently not
sure how to best avoid it without "punching through" the error
layers I've built up.
