+++ llvm-mapper/src/block/module.rs
@@ -6,7 +6,7 @@ use crate::block::attributes::{AttributeGroups, Attributes};
 use crate::block::type_table::TypeTable;
 use crate::block::{BlockId, BlockMapError, IrBlock};
 use crate::map::{MapCtx, Mappable};
-use crate::record::{Comdat, DataLayout};
+use crate::record::{Comdat, DataLayout, RecordMapError};
 use crate::unroll::UnrolledBlock;
 
 /// Models the `MODULE_BLOCK` block.
@@ -41,7 +41,7 @@ impl IrBlock for Module {
 
         // Each module *should* have a target triple, but doesn't necessarily.
         let triple = if let Some(record) = block.maybe_one_record(ModuleCode::Triple as u64)? {
-            record.try_string(0)?
+            record.try_string(0).map_err(RecordMapError::from)?
         } else {
             TARGET_TRIPLE.into()
         };
@@ -49,7 +49,7 @@ impl IrBlock for Module {
         // Each module *should* have a datalayout record, but doesn't necessarily.
         let datalayout =
             if let Some(record) = block.maybe_one_record(ModuleCode::DataLayout as u64)? {
-                DataLayout::try_map(record, ctx)?
+                DataLayout::try_map(record, ctx).map_err(RecordMapError::from)?
             } else {
                 DataLayout::default()
             };
@@ -57,7 +57,8 @@ impl IrBlock for Module {
         // Each module has zero or exactly one MODULE_CODE_ASM records.
         let asm = match block.maybe_one_record(ModuleCode::Asm as u64)? {
             Some(rec) => rec
-                .try_string(0)?
+                .try_string(0)
+                .map_err(RecordMapError::from)?
                 .split('\n')
                 .map(String::from)
                 .collect::<Vec<_>>(),
@@ -68,25 +69,29 @@ impl IrBlock for Module {
         let deplibs = block
             .records(ModuleCode::DepLib as u64)
             .map(|rec| rec.try_string(0))
-            .collect::<Result<Vec<_>, _>>()?;
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(RecordMapError::from)?;
 
         // Build the section table. We'll reference this later.
         let _section_table = block
             .records(ModuleCode::SectionName as u64)
             .map(|rec| rec.try_string(0))
-            .collect::<Result<Vec<_>, _>>()?;
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(RecordMapError::from)?;
 
         // Build the GC table. We'll reference this later.
         let _gc_table = block
             .records(ModuleCode::GcName as u64)
             .map(|rec| rec.try_string(0))
-            .collect::<Result<Vec<_>, _>>()?;
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(RecordMapError::from)?;
 
         // Build the Comdat list. We'll reference this later.
         let _comdats = block
             .records(ModuleCode::Comdat as u64)
             .map(|rec| Comdat::try_map(rec, ctx))
-            .collect::<Result<Vec<_>, _>>()?;
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(RecordMapError::from)?;
 
         // Build the type table.
         let type_table = TypeTable::try_map(block.one_block(BlockId::Ir(IrBlockId::Type))?, ctx)?;
