+++ llvm-mapper/src/record/mod.rs
@@ -4,18 +4,33 @@
 //! structure or mapping implementation. Simpler records are mapped inline within their
 //! blocks.
 
+pub mod comdat;
 pub mod datalayout;
 
-use std::convert::TryInto;
+use std::num::TryFromIntError;
+use std::string::FromUtf8Error;
 
-use llvm_support::StrtabRef;
-use num_enum::TryFromPrimitive;
 use thiserror::Error;
 
+pub use self::comdat::*;
 pub use self::datalayout::*;
 use crate::block::StrtabError;
-use crate::map::{MapCtx, MapCtxError, Mappable};
-use crate::unroll::UnrolledRecord;
+use crate::map::{MapCtxError, Mappable};
+
+/// Potential errors when trying to extract a string from a record.
+#[non_exhaustive]
+#[derive(Debug, Error)]
+pub enum RecordStringError {
+    /// The start index for the string is invalid.
+    #[error("impossible string index: {0} >= {1} (field count)")]
+    BadIndex(usize, usize),
+    /// A field in the record is too large to fit in a byte.
+    #[error("impossible character value in string: {0}")]
+    BadCharacter(#[from] TryFromIntError),
+    /// The string doesn't look like valid UTF-8.
+    #[error("invalid string encoding: {0}")]
+    BadEncoding(#[from] FromUtf8Error),
+}
 
 /// Potential errors when mapping a single bitstream record.
 #[non_exhaustive]
@@ -33,6 +48,10 @@ pub enum RecordMapError {
     #[error("error while mapping record: {0}")]
     BadRecordLayout(String),
 
+    /// We encountered a string we couldn't parse.
+    #[error("error while parsing string: {0}")]
+    BadRecordString(#[from] RecordStringError),
+
     /// Our mapping context was invalid for our operation.
     #[error("invalid mapping context: {0}")]
     BadContext(#[from] MapCtxError),
@@ -44,67 +63,4 @@ pub enum RecordMapError {
     /// We encountered an unsupported feature or layout.
     #[error("unsupported: {0}")]
     Unsupported(String),
-}
-
-/// The different kinds of COMDAT selections.
-///
-/// This is a nearly direct copy of LLVM's `SelectionKind`; see `IR/Comdat.h`.
-#[non_exhaustive]
-#[derive(Debug, TryFromPrimitive)]
-#[repr(u64)]
-pub enum ComdatSelectionKind {
-    /// The linker may choose any COMDAT.
-    Any,
-    /// The data referenced by the COMDAT must be the same.
-    ExactMatch,
-    /// The linker will choose the largest COMDAT.
-    Largest,
-    /// No deduplication is performed.
-    NoDeduplicate,
-    /// The data referenced by the COMDAT must be the same size.
-    SameSize,
-}
-
-/// Models the `MODULE_CODE_COMDAT` record.
-#[non_exhaustive]
-#[derive(Debug)]
-pub struct Comdat {
-    /// The selection kind for this COMDAT.
-    pub selection_kind: ComdatSelectionKind,
-    /// The COMDAT key.
-    pub name: String,
-}
-
-impl Mappable<UnrolledRecord> for Comdat {
-    type Error = RecordMapError;
-
-    fn try_map(record: &UnrolledRecord, ctx: &mut MapCtx) -> Result<Self, Self::Error> {
-        if !ctx.use_strtab()? {
-            return Err(RecordMapError::Unsupported(
-                "v1 COMDAT records are not supported".into(),
-            ));
-        }
-
-        // v2: [strtab offset, strtab size, selection kind]
-        if record.fields().len() != 3 {
-            return Err(RecordMapError::BadRecordLayout(format!(
-                "expected exactly 3 fields in COMDAT record, got {}",
-                record.fields().len()
-            )));
-        }
-
-        // Index safety: we check for at least 3 fields above.
-        let name = {
-            let sref: StrtabRef = (record.fields()[0], record.fields()[1]).into();
-            ctx.strtab()?.try_get(&sref)?
-        };
-        let selection_kind: ComdatSelectionKind = record.fields()[2].try_into().map_err(|e| {
-            RecordMapError::BadRecordLayout(format!("invalid COMDAT selection kind: {:?}", e))
-        })?;
-
-        Ok(Self {
-            selection_kind,
-            name: name.into(),
-        })
-    }
 }
